Some tile memory optimization

+++ SF3Lib/Models/Files/MPD/SurfaceTile.FileOnly.cs
@@ -5,7 +5,7 @@ namespace SF3.Models.Files.MPD {
         public float GetSurfaceDataVertexHeight(CornerType corner)
             => MPD_File.SurfaceDataChunk?.HeightmapRowTable?[Y]?.GetHeight(X, corner) ?? 0.0f;
         public float GetSurfaceModelVertexHeight(CornerType corner) {
-            var bl = _blockVertexLocations[corner];
+            var bl = _blockVertexLocations[(int) corner];
             return ((MPD_File.SurfaceModelChunk?.VertexHeightBlockTable?[bl.Num]?[bl.X, bl.Y]) ?? 0) / 16f;
         }
     }
